Implementation group and a bit fix user

+++ webApi/src/main/java/com/edu/moneywayapi/webApi/controller/GroupController.java
@@ -1,23 +1,32 @@
 package com.edu.moneywayapi.webApi.controller;
 
 import br.com.fluentvalidator.context.ValidationResult;
+import com.edu.moneywayapi.domain.entity.Group;
+import com.edu.moneywayapi.domain.entity.User;
 import com.edu.moneywayapi.domain.exception.NoSuchGroupException;
-import com.edu.moneywayapi.domain.exception.NoSuchUserException;
 import com.edu.moneywayapi.domain.service.GroupService;
+import com.edu.moneywayapi.domain.service.UserService;
 import com.edu.moneywayapi.webApi.dto.GroupDTO;
 import com.edu.moneywayapi.webApi.mapper.GroupDTOMapper;
 import com.edu.moneywayapi.webApi.validator.GroupValidator;
+import com.jayway.jsonpath.JsonPath;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
+import java.security.Principal;
+import java.util.List;
+
 @RestController
-@RequestMapping(value = "/group")
+@RequestMapping(value = "/groups")
 public class GroupController {
 
     @Autowired
     private GroupService groupService;
+
+    @Autowired
+    private UserService userService;
 
     @Autowired
     private GroupValidator groupValidator;
@@ -26,7 +35,10 @@ public class GroupController {
     private GroupDTOMapper groupDTOMapper;
 
     @GetMapping("/{id}")
-    public ResponseEntity<?> getById(@PathVariable(name = "id") Long id) {
+    public ResponseEntity<?> getById(Principal principal, @PathVariable(name = "id") Long id) {
+        if (!groupService.existsUser(id, principal.getName()))
+            return new ResponseEntity<>(HttpStatus.FORBIDDEN);
+
         try {
             GroupDTO group = groupDTOMapper.map(groupService.findById(id));
             return new ResponseEntity<>(group, HttpStatus.OK);
@@ -35,8 +47,12 @@ public class GroupController {
         }
     }
 
-    @GetMapping("/{token}")
-    public ResponseEntity<?> getByToken(@PathVariable(name = "token") String token) {
+    @GetMapping()
+    public ResponseEntity<?> getByToken(@RequestBody String requestJson) {
+        String token = JsonPath.read(requestJson, "$.token");
+        if (token == null)
+            return new ResponseEntity<>("Токен отсутствует", HttpStatus.BAD_REQUEST);
+
         try {
             GroupDTO group = groupDTOMapper.map(groupService.findByToken(token));
             return new ResponseEntity<>(group, HttpStatus.OK);
@@ -55,8 +71,12 @@ public class GroupController {
         return new ResponseEntity<>(HttpStatus.CREATED);
     }
 
-    @GetMapping("/delete/{id}")
-    public ResponseEntity<?> deleteById(@PathVariable(name = "id") Long id) {
+    @DeleteMapping("/{id}")
+    public ResponseEntity<?> deleteById(Principal principal, @PathVariable(name = "id") Long id) {
+        User user = userService.findByLogin(principal.getName());
+        if (!groupService.isOwner(id, user.getId()))
+            return new ResponseEntity<>(HttpStatus.FORBIDDEN);
+
         try {
             groupService.deleteById(id);
             return new ResponseEntity<>(HttpStatus.OK);
@@ -66,8 +86,45 @@ public class GroupController {
     }
 
     @DeleteMapping("/{id}/user")
-    public ResponseEntity<?> deleteUser(@PathVariable Long id, @RequestParam String userLogin) {
+    public ResponseEntity<?> deleteUser(Principal principal, @PathVariable Long id, @RequestBody String requestJson) {
+        User user = userService.findByLogin(principal.getName());
+        if (!groupService.isOwner(id, user.getId()))
+            return new ResponseEntity<>(HttpStatus.FORBIDDEN);
+
+        String userLogin = JsonPath.read(requestJson, "$.userLogin");
+        if (userLogin == null)
+            return new ResponseEntity<>("Логин пользователя отсутствует", HttpStatus.BAD_REQUEST);
+
+        if (!groupService.existsUser(id, userLogin))
+            return new ResponseEntity<>("Пользователь не найден", HttpStatus.BAD_REQUEST);
+
         groupService.deleteUser(id, userLogin);
         return new ResponseEntity<>(HttpStatus.OK);
+    }
+
+    @PostMapping("/{id}/user")
+    public ResponseEntity<?> addUser(Principal principal, @PathVariable Long id) {
+        if (!groupService.existsUser(id, principal.getName()))
+            return new ResponseEntity<>("Пользователь уже в группе", HttpStatus.CONFLICT);
+
+        User user = userService.findByLogin(principal.getName());
+        groupService.addUser(id, user.getId());
+        return new ResponseEntity<>(HttpStatus.OK);
+    }
+
+    @GetMapping("/user")
+    public ResponseEntity<?> getByUser(Principal principal) {
+        User user = userService.findByLogin(principal.getName());
+        List<Group> groups = user.getGroups();
+        return new ResponseEntity<>(groups, HttpStatus.OK);
+    }
+
+    @GetMapping("/{id}/users")
+    public ResponseEntity<?> getUsers(Principal principal, @PathVariable Long id) {
+        if (!groupService.existsUser(id, principal.getName()))
+            return new ResponseEntity<>(HttpStatus.FORBIDDEN);
+
+        List<String> users = groupService.getUsers(id);
+        return new ResponseEntity<>(users, HttpStatus.OK);
     }
 }
